Move memory scanner class to its own file
